refactor: :recycle: refactor of getProjects
