+++ src/cmsFaker/projects/variables.ts
@@ -1,0 +1,2 @@
+export const SEO_GITHUB = "github",
+  SEO_LINK = "link"
