+++ src/components/organisms/ProjectContent.tsx
@@ -1,7 +1,7 @@
 import { FC } from "react";
 import styled, { css } from "styled-components";
 import { Container } from "../../../styles/system/Container";
-import { IProjects } from "../../cmsFaker/getProjects";
+import { IProjects } from "../../cmsFaker/projects/interfaces";
 import { getResumeProject } from "../../helper/getResumeProject";
 import { PropStyleTheme } from "../../types";
 import { ProjectCard } from "../molecules/ProjectCard";
